// src/app/common/enums/AccessibilityButton.enum.ts
export enum AccessibilityButtonEnum {
  NormalFont = 'Normal font',
  BiggerFont = 'Bigger font',
  BiggestFont = 'Biggest Font',
  BoldFont = 'BoldFont',
  Hyperlink = 'Hyperlink',
  ResetContrast = 'ResetContrast',
  BlackWhite = 'Black White',
  WhiteBlack = 'White Black',
  BlackYellow = 'Black Yellow',
  YellowBlack = 'Yellow Black'
}